Variant factory in process

+++ include/network/common/trans/receive.h
@@ -7,7 +7,7 @@
 #include "sys/error_print.h"
 #include "network/common/def.h"
 namespace network{
-    ErrorCode receive(Socket socket,std::ranges::random_access_range auto& buffer, size_t n) noexcept{
+    ErrorCode receive(Socket socket,std::ranges::random_access_range auto& buffer, uint64_t n) noexcept{
         if(n==0)
             return ErrorCode::NONE;
         ssize_t read_num = recv(socket,buffer.data(),n,MSG_WAITALL);
